+++ src/confusius/plotting/__init__.py
@@ -8,6 +8,7 @@ __all__ = [
     "plot_contours",
     "plot_matrix",
     "plot_napari",
+    "plot_stat_map",
     "plot_volume",
     "VolumePlotter",
 ]
@@ -17,6 +18,7 @@ from confusius.plotting.image import (
     plot_carpet,
     plot_composite,
     plot_contours,
+    plot_stat_map,
     plot_volume,
 )
 from confusius.plotting.matrix import plot_matrix